+++ src/types/index.ts
@@ -12,3 +12,4 @@ export * from "./linkage";
 export * from "./macro";
 export * from "./adapter";
 export * from "./filters";
+export * from "./calc";